Move seed functionality to own module
